Automatically add imported transactions into the sql database
Split SqlRepository into AbstractSqlRepository and TransactionSqlRepository, so that in the future I can implement CategorySqlRepository
Add Category model that will will be used in classification.
Add a base clasr AbstractClassifier that will be used to categorise the transactions.
Fixed AbstractSqlRepository initialisation method to handle multiple statements in the configuration file.